refactored, but running into circular import issues
