+++ quixstreams/models/topics/exceptions.py
@@ -1,0 +1,13 @@
+from quixstreams.exceptions import QuixException
+
+
+class MissingTopicAdmin(QuixException):
+    ...
+
+
+class TopicValidationError(QuixException):
+    ...
+
+
+class MissingTopicForChangelog(QuixException):
+    ...
