+++ quixstreams/platforms/quix/__init__.py
@@ -1,3 +1,4 @@
 from .api import *
-from .config import *
 from .checks import *
+from .config import *
+from .topic_manager import *
