+++ tests/test_quixstreams/test_models/test_topics/test_topics.py
@@ -27,7 +27,7 @@ from quixstreams.models.serializers import (
     SERIALIZERS,
     DESERIALIZERS,
 )
-from .utils import ConfluentKafkaMessageStub, int_to_bytes, float_to_bytes
+from ..utils import ConfluentKafkaMessageStub, int_to_bytes, float_to_bytes
 
 
 class JSONListDeserializer(JSONDeserializer):
